Implement river rendering on the GPU
* Render each map triangle with a GPU triangle
* Draw two river textures, one for river forks and one for river bends
* Generate geometry that writes the u,v textures depending on whether
  it's a fork or bend

Unfortunately the river sizes didn't work as well as I had hoped.
There's a wide variety of triangle sizes, and I need to take them into
account. I also think it may be useful to explicitly construct the
mipmaps to not exactly match the larger texture but instead make the
pixel blue if *any* of the larger mipmap's corresponding pixels are
blue. That might prevent the rivers from becoming too thin and
disappearing. Or maybe I can solve it in some other way.

There's also another approach to river drawing that I want to try out.