Optimize static mountain calculation
The experiment with mountain placement was successful, but the code was
rather slow. It looped through all mountains and calculated the
distance to the closest one.

The new code inverts this.

It goes through all mountains and calculates a distance field, similar
to Worley Noise in spirit but different in implementation.

The final elevation will be a *mix* of the base elevation and the
elevation depending on the distance to the closest mountain. It's
roughly 10X as fast as the old experimental code.