Change elevation painting to use feathered brush
Previous system felt unintuitive at times. The main problem was that
the constraints set the elevation, and then the user sees the
elevation and wants to change it. But since the constraints that
caused that elevation might be elsewhere, painting on top of the
output elevation doesn't necessarily change what you want.

The new system doesn't use constraints. Instead, paint the elevation
directly. The brushes raise and lower the elevation. There's a
gradient from the center of the brush to the outer edge where the
effect is less.

* Remove distance field calculation, as it was only needed to turn
  constraints into elevation. This was a fairly expensive part of the
  map generation (roughly 20-30%) so removing it sped things up.
* Cache the noise values used for valley elevation. This was also a
  fairly expensive part of the map generation (roughly 15%), and I had
  put it in to see if I liked the effect. I decided I like the effect
  and want to keep it, so I moved it to the precomputation function.
* Remove r_ocean. It's no longer needed.
* Add bilinear interpolation for points sampled from the painted
  elevation. I didn't need this previously because the graph distances
  hid the grid nature of the painted minimap, but after switching to
  elevation painting, the grid artifacts were obvious. Bilinear
  interpolation hides them.