Change elevation painting to use entire stroke for feathering
Problem: when drawing repeatedly over an area, the feathered brush
turns into a solid brush.

Solution: change the way feathering works to act at the *stroke*
level. When starting a brush stroke, store the previous elevation
and reset the brush strength to 0. Then apply the brush effect to
the brush strength array, which is capped at the current brush
strength.

The brush effect is based on the amount of time since the last mouse
move. This creates a constant rate (but capped to allow for pauses or
lag). It varies by size of brush, with smaller brushes having a higher
rate.

If you repeatedly paint over the same area, the brush strength array
maxes out at the brush strength, so feathered areas don't become
solid.

If you repeatedly paint over an area with the feathered edge of the
brush, and then paint with the center, it won't suddenly become
strong, but instead take time just as if you had started painting
there with the center.