Workaround for bug: rivers weren't reaching the ocean
The root cause of this is that the river rendering depends on knowing
where the water flows. Since coastal triangles were the seed for the
Dijkstra search, there's no outflow from them, and they didn't get
rendered. The long term fix is to generate special river delta sprites
for coastal triangles, or alternatively, set the outflow for them.

* Add workaround: make dijkstra_search start with oceans as the seed
  points, so that they flow through the coastal triangles and set the
  outflow pointer.
* Fix potential bug: make river geometry buffer big enough to support
  1.5 output triangles per input triangle, since each fork can produce
  2 output triangles. The worst case size is from a maximally filled
  binary tree. If there are N/2 leaves with N/2 nodes, that's N/2
  triangles + 2 * N/2 triangles = 3/2*N triangles.
* Change dijkstra_search to never expand in oceans.
* Fix bug: disallow 0 painted elevation, to ensure that coastal
  triangles are *only* produced by algorithm and never from input data.
* Remove unused elevation() function; this is now performed in painting.js.