Add download button for Safari (desktop, ios)
Right click to save only worked in Chrome/desktop. The download
button works by reading pixels from WebGL and copying them to a
Canvas, reversing the rows along the way.

I don't have this working across browsers. In Firefox/android and also
Firefox/desktop, it successfully copies to a canvas, but the "click"
action doesn't seem to trigger a download, probably because the
screenshot has to be rendered later and it's no longer associated with
the user action. Not sure. In Chrome/android, it's unable to copy to a
canvas, even when I use preserveDrawingBuffer. Not sure what's wrong.