Feat: move buttons+sliders to bottom in portrait mode
I had previously adjusted the sliders to work for Firefox in landscape
mode, but that made things worse in portrait mode. I decided to change
the portrait mode layout to put the map up top and the controls
underneath. This is primarily for touch devices where your hand is
near the bottom of the screen, so moving a slider at the top of the
screen would cover up the map. Controls at the bottom of the screen
allow you to see the result without your hand covering it.

* Made sliders take up multiple columns, since there's typically more
  space in portrait mode. No need to scroll through them.
* Added a highlight outline around the current tool.