Feat: smaller pen, variable pen size with Apple Pencil, slow draw
Features: hold down shift to paint more slowly.

Added a smaller brush for fine detail work.

On ipad, Apple Pencil is pressure sensitive, so pushing harder will
now draw larger features.

To make the Apple Pencil feature work, fixed a bug in draggable.js that
was inconsistent in how it passed the raw event to the caller. For
mouse it was using the operation object for the MouseEvent and for
touch it was using the position object for the Touch object but not
the TouchEvent object. Now it always uses the position object, and
puts the event (MouseEvent/TouchEvent) in and also puts the Touch in
if it's a touch event.