Refactor: move dual-mesh's ghost construction into the core
The idea is that *point selection* and *boundary points* can vary
between projects, but *ghost construction* is going to be the same for
(almost) every project. Some projects might not use ghost
construction, but if they do, they'll all use the same algorithm, so
that will go in the core library.