Refactor: separate input and output dual-mesh types
Previously, a TriangleMesh could be in various states of being
constructed. Before and after construction the types looked very
similar. With this change, MeshInitializer is the "before"
construction and TriangleMesh is the "after" construction. They look
different and can be distinguished both by their type and by their run
time structure.

Boundary points and ghost structures are added at the MeshInitializer
stage, working directly with the Delaunator data. The indices and
accessors are added at the TriangleMesh stage, wrapping the Delaunator
data with accessors.