Refactor: add boundary generators to dual-mesh; remove MeshBuilder
There's a tradeoff between dual-mesh's ease of use and flexibility. The
MeshBuilder was trying to be easy to use, but over time the various
use cases didn't fit into it, so it became more convoluted. I decided
to remove MeshBuilder, as it's not that useful for mapgen4 or mapgen2.
Instead, dual-mesh exposes the boundary point generators and then the
map generator can call those functions.