Feat: travelers calculate movement when switching region, not ahead of time
I had been calculating the movement costs at pathfinding time, but
that means if we change the map, we're still using the original
movement costs instead of the new costs. Changed it to calculate the
movement costs when calculating the next line segment for animation.
This still isn't as responsive as calculating every frame though.