Refactor: switch render.ts from regl.js to using WebGL directly
My main motivation is that I'd like to use WebGL 2 features in other
projects that use mapgen4's renderer, and regl doesn't fully support
WebGL 2:

  * https://github.com/regl-project/regl/issues/378
  * https://github.com/regl-project/regl/issues/561
  * https://github.com/regl-project/regl/issues/667
    "regl doesn't (and won't) aim to address WebGL 2"

But secondarily, regl.js is half the bundle size, so removing it will
make the page load faster.

The goal is NOT to change the rendering in mapgen4, but I do want to
change it in mapgen4-adjacent projects. In mapgen4 this renderer
generates the same output as the previous one.

The new version is roughly the same amount of code as the old. I
cleaned up some of the code while doing this refactor.