Feat: use a gpu shader instead of cpu canvas to draw river shapes
Instead of prerendering many different river widths, we can generate
the different river widths in the shader. This is less code *and* the
rivers look sharper.

However *it does not look the same* so I adjusted some of the rendering
parameters to try to compensate. It still doesn't look the same. And
it could be better, but that's a future project.