Feat: use WebGL R16F format to store elevation data in renderer
In WebGL1 I used separate R,G channels to store the low and high bytes
of the elevation. This led to some visual artifacts, and also
prevented using linear texture filtering.

I switched this to use a single 16-bit float channel. It improves the
output quality. But some of the visual artifacts looked interesting,
so I might need to add that back intentionally instead of relying on
the accidental artifacts.